Don't need to convert to int from Integer for AppdefEntityID

+++ src/org/hyperic/hq/appdef/shared/AppdefGroupValue.java
@@ -6,7 +6,7 @@
  * normal use of the program, and does *not* fall under the heading of
  * "derived work".
  * 
- * Copyright (C) [2004, 2005, 2006], Hyperic, Inc.
+ * Copyright (C) [2004-2008], Hyperic, Inc.
  * This file is part of HQ.
  * 
  * HQ is free software; you can redistribute it and/or modify
@@ -354,8 +354,7 @@ public class AppdefGroupValue
     // utility method for converting back and forth from appdef to authz
     private AppdefEntityID entryToEntity (GroupEntry entry) {
         return new AppdefEntityID ( AppdefUtil.resNameToAppdefTypeId(
-                                    entry.getType()),
-                                    entry.getId().intValue());
+                                    entry.getType()), entry.getId());
     }
